// ImageProperties.d.ts
/**
 * ImageProperties
 * @properties for Image
 */
interface ImageProperties {
    backgroundColor?: string;
    backgroundImage?: string;
    color?: string;
    fontFamily?: string;
    fontSize?: number;
    height?: number;
    width?: number;
    padding?: number;
    textAlign?: string;
    wordWrap?: string;
    display?: string;
    verticalAlign?: string;
    boxSizing?: string;
    backgroundSize?: string;
    fontStyle?: string;
    textDecoration?: string;
}
export default ImageProperties;